Merge a bunch of changes into master

Approved-by: Cameron Rivers
Approved-by: Troy Shanks

+++ c_library/gamelink_c.h
@@ -1,11 +1,11 @@
 #pragma once
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
-
-	// x_Make always requires a x_Kill call or you will have memory leaks! 
+	// x_Make always requires a x_Kill call or you will have memory leaks!
 	// x_Init does not require a x_Kill call
 	// Example: MuxyGameLink_Schema_DatastreamEvent_MakeJson requires a call to MuxyGameLink_Schema_DatastreamEvent_KillJson
 	// If it were named MuxyGameLink_Schema_DatastreamEvent_InitJson it wouldn't require a kill call
@@ -57,10 +57,11 @@ extern "C"
 	void MuxyGameLink_Kill(MuxyGameLink GameLink);
 
 	void MuxyGameLink_ForeachPayload(MuxyGameLink GameLink, void (*Callback)(void* UserData, MGL_Payload Payload), void* UserData);
-	size_t MuxyGameLink_Payload_GetSize(MGL_Payload Payload);
+	uint32_t MuxyGameLink_Payload_GetSize(MGL_Payload Payload);
 	const char* MuxyGameLink_Payload_GetData(MGL_Payload Payload);
 
-	// GetFirstError may not need MGL_SCHEMA_RESPONSE_TYPE in later versions of the library, as we will internally tag structs with the RespType so the burden isn't on the caller
+	// GetFirstError may not need MGL_SCHEMA_RESPONSE_TYPE in later versions of the library, as we will internally tag structs with the
+	// RespType so the burden isn't on the caller
 	MGL_Error MuxyGameLink_Schema_GetFirstError(void* Resp, MGL_SCHEMA_RESPONSE_TYPE RespType);
 	bool MuxyGameLink_Error_IsValid(MGL_Error Error);
 	uint32_t MuxyGameLink_Error_GetCode(MGL_Error Error);
@@ -70,16 +71,16 @@ extern "C"
 	bool MuxyGameLink_ReceiveMessage(MuxyGameLink GameLink, const char* Bytes, uint32_t Length);
 
 	MGL_RequestId MuxyGameLink_AuthenticateWithPIN(MuxyGameLink GameLink,
-							const char* ClientId,
-							const char* PIN,
-							void (*Callback)(void* UserData, MGL_Schema_AuthenticateResponse AuthResp),
-							void* UserData);
+												   const char* ClientId,
+												   const char* PIN,
+												   void (*Callback)(void* UserData, MGL_Schema_AuthenticateResponse AuthResp),
+												   void* UserData);
 
 	MGL_RequestId MuxyGameLink_AuthenticateWithRefreshToken(MuxyGameLink GameLink,
-							const char* ClientId,
-							const char* RefreshToken,
-							void (*Callback)(void* UserData, MGL_Schema_AuthenticateResponse AuthResp),
-							void* UserData);
+															const char* ClientId,
+															const char* RefreshToken,
+															void (*Callback)(void* UserData, MGL_Schema_AuthenticateResponse AuthResp),
+															void* UserData);
 
 	bool MuxyGameLink_IsAuthenticated(MuxyGameLink GameLink);
 
@@ -92,13 +93,16 @@ extern "C"
 
 	MGL_RequestId MuxyGameLink_SubscribeToDatastream(MuxyGameLink GameLink);
 	MGL_RequestId MuxyGameLink_UnsubscribeFromDatastream(MuxyGameLink GameLink);
+	// Do not store MGL_Schema_DatastreamEvent's or MGL_Schema_DatastreamUpdate's anywhere, they may be invalidated at the moment the
+	// callback finishes.
 	uint32_t MuxyGameLink_OnDatastream(MuxyGameLink GameLink,
-					void (*Callback)(void* UserData, MGL_Schema_DatastreamUpdate DatastreamUpdate),
-					void* UserData);
+									   void (*Callback)(void* UserData, MGL_Schema_DatastreamUpdate DatastreamUpdate),
+									   void* UserData);
 	void MuxyGameLink_DetachOnDatastream(MuxyGameLink GameLink, uint32_t OnDatastreamHandle);
 
-	size_t MuxyGameLink_Schema_DatastreamUpdate_GetEventCount(MGL_Schema_DatastreamUpdate DatastreamUpdate);
-	MGL_Schema_DatastreamEvent MuxyGameLink_Schema_DatastreamUpdate_GetEventAt(MGL_Schema_DatastreamUpdate DatastreamUpdate, size_t AtIndex);
+	uint32_t MuxyGameLink_Schema_DatastreamUpdate_GetEventCount(MGL_Schema_DatastreamUpdate DatastreamUpdate);
+	MGL_Schema_DatastreamEvent MuxyGameLink_Schema_DatastreamUpdate_GetEventAt(MGL_Schema_DatastreamUpdate DatastreamUpdate,
+																			   uint32_t AtIndex);
 
 	char* MuxyGameLink_Schema_DatastreamEvent_MakeJson(MGL_Schema_DatastreamEvent DatastreamEvent);
 	void MuxyGameLink_Schema_DatastreamEvent_KillJson(char* DatastreamEventJson);
@@ -109,8 +113,8 @@ extern "C"
 	MGL_RequestId MuxyGameLink_SubscribeToAllPurchases(MuxyGameLink GameLink);
 	MGL_RequestId MuxyGameLink_UnsubscribeFromAllPurchases(MuxyGameLink GameLink);
 	uint32_t MuxyGameLink_OnTwitchPurchaseBits(MuxyGameLink GameLink,
-						void (*Callback)(void* UserData, MGL_Schema_TwitchPurchaseBitsResponse TPBResp),
-						void* UserData);
+											   void (*Callback)(void* UserData, MGL_Schema_TwitchPurchaseBitsResponse TPBResp),
+											   void* UserData);
 	void MuxyGameLink_DetachOnTwitchPurchaseBits(MuxyGameLink GameLink, uint32_t id);
 
 	const char* MuxyGameLink_Schema_TwitchPurchaseBits_GetId(MGL_Schema_TwitchPurchaseBitsResponse TPBResp);
